stages/05-dashboard-build/output/reports/requirements.txt
supabase==2.7.4
python-dotenv==1.0.1
jinja2==3.1.4
pdfkit==1.0.0
resend==2.5.0
pandas==2.2.3
pytest==8.3.3
